Fix: Address review comments
