+++ VideoGameDatabase/VideoGameDatabase/Models/GameDetailClip.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+struct GameDetailClipResponse: Decodable {
+    let clip: GameClip?
+}
+struct GameClip: Decodable {
+    let clip: String?
+    let clips: [String: String]?
+}
